Rename custom types module to avoid name collision
This was stopping docker dev env from starting correctly for some reason.
